agent-tool-calling/pyproject.toml
[project]
name = "agent-tool-calling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.6.2",
    "openai-agents>=0.2.1",
    "python-dotenv>=1.1.1",
    "selenium>=4.34.2",
    "twilio>=9.6.5",
    "webdriver-manager>=4.0.2",
]
